Continued work on Grouping tab. Squashed a couple of bugs.
Fixed an error in labelBox that was keeping it from properly parsing options for expansion of widgets.
Fixed an error in tryLoadDesc that was clobbering each item as the next was found.
Added buttons for adding and subtracting items to a queue for inclusion in RIG file.
Started working on load function for grouping files.