Improvements to PickEd (Options, etc)
Fixed error causing end date to be skipped when processing days in range.
Added Options dialog file from other projects to this one, giving users ability to set options through the GUI.
Added an option to the hash set for insertDateWidget, allowing background of date widget to be set.
Filled out a mock-up of GUI elements that'll be coded in later commits.
Fixed a few typos.